Moved SessionTimeOut value from SYSTEM_CONFIGURATION table to CONFIG_KEY_VALUE_INTEGER table.  Updated Mayfly jar with changes to support ATLTER TABLE...INSERT AFTER. Some code cleanup.
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@11598 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/org/mifos/framework/components/configuration/business/Configuration.java
@@ -75,7 +75,7 @@ public class Configuration {
   private void initializeSystemConfiguration(){
 	 systemConfig = new SystemConfiguration((SupportedLocalesEntity)cacheRepo.getValueFromSystemCache(ConfigConstants.MFI_LOCALE),
 			  (MifosCurrency)cacheRepo.getValueFromSystemCache(ConfigConstants.CURRENCY),
-			  (Short)cacheRepo.getValueFromSystemCache(ConfigConstants.SESSION_TIMEOUT),
+			  (Integer)cacheRepo.getValueFromSystemCache(ConfigConstants.SESSION_TIMEOUT),
 			  (Integer)cacheRepo.getValueFromSystemCache(ConfigConstants.TIMEZONE));
   }
